# assets/swagger/swagger.v1.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This documentation describes the RTME/CFG API.",
    "title": "V1 API.",
    "version": "1.0.0"
  },
  "basePath": "/api/v1",
  "paths": {
    "/app/config": {
      "post": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "app"
        ],
        "summary": "Setup app config",
        "operationId": "appConfig",
        "parameters": [
          {
            "type": "string",
            "description": "db type",
            "name": "DBTYPE",
            "in": "body",
            "required": true
          },
          {
            "type": "string",
            "description": "db rtme url",
            "name": "DBRTMEURL",
            "in": "body",
            "required": true
          },
          {
            "type": "string",
            "description": "db cfg url",
            "name": "DBCFGURL",
            "in": "body",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/app/status": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "app"
        ],
        "summary": "Get app status",
        "operationId": "appStatus",
        "responses": {
          "200": {
            "description": "status response"
          }
        }
      }
    },
    "/cfg/groups": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "config"
        ],
        "summary": "Get group name list",
        "operationId": "cfgGroups",
        "responses": {
          "200": {
            "description": "Groups names"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/cfg/groups/{name}/users": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "config"
        ],
        "summary": "Get group agent list",
        "operationId": "cfgGroups",
        "parameters": [
          {
            "type": "string",
            "description": "name of group",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Agent dbid list"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/cfg/places/{dbid}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "config"
        ],
        "summary": "Get info on a place",
        "operationId": "cfgPlace",
        "parameters": [
          {
            "type": "string",
            "description": "id of place",
            "name": "dbid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Place info"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/rtme/{date}/login": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "rtme"
        ],
        "summary": "Export rtme info of login.",
        "operationId": "rtmeLogin",
        "parameters": [
          {
            "type": "string",
            "description": "date",
            "name": "date",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "filter by group/instance name",
            "name": "group",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Logins info",
            "$ref": "#/responses/RTMELoginResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/rtme/{date}/status": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "rtme"
        ],
        "summary": "Export rtme info of status.",
        "operationId": "rtmeStatus",
        "parameters": [
          {
            "type": "string",
            "description": "date",
            "name": "date",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "filter by group/instance name",
            "name": "group",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Status info",
            "$ref": "#/responses/RTMEStatusResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    }
  },
  "definitions": {
    "Error": {
      "description": "Error is error format response",
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-package": "github.com/sapk/rtme-browser/router/api/common"
    },
    "FormattedLoginResp": {
      "description": "FormattedLoginResp format the login to be used by api",
      "type": "object",
      "properties": {
        "End": {
          "type": "integer",
          "format": "int64"
        },
        "Sessions": {
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Session"
            }
          }
        },
        "Start": {
          "type": "integer",
          "format": "int64"
        },
        "Users": {
          "type": "object",
          "additionalProperties": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "x-go-package": "github.com/sapk/rtme-browser/vendor/github.com/sapk/go-genesys/db/rtme"
    },
    "FormattedStatusResp": {
      "description": "FormattedStatusResp format the status to be used by api",
      "type": "object",
      "properties": {
        "End": {
          "type": "integer",
          "format": "int64"
        },
        "Start": {
          "type": "integer",
          "format": "int64"
        },
        "Status": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StatusEvent"
          }
        }
      },
      "x-go-package": "github.com/sapk/rtme-browser/vendor/github.com/sapk/go-genesys/db/rtme"
    },
    "Session": {
      "description": "Session a login session",
      "type": "object",
      "properties": {
        "End": {
          "type": "integer",
          "format": "int64"
        },
        "Place": {
          "type": "integer",
          "format": "int64"
        },
        "Start": {
          "type": "integer",
          "format": "int64"
        }
      },
      "x-go-package": "github.com/sapk/rtme-browser/vendor/github.com/sapk/go-genesys/db/rtme"
    },
    "StatusEvent": {
      "description": "StatusEvent a status event",
      "type": "object",
      "properties": {
        "Agent": {
          "type": "integer",
          "format": "int64"
        },
        "End": {
          "type": "integer",
          "format": "int64"
        },
        "Place": {
          "type": "integer",
          "format": "int64"
        },
        "Start": {
          "type": "integer",
          "format": "int64"
        },
        "Status": {
          "type": "integer",
          "format": "int64"
        }
      },
      "x-go-package": "github.com/sapk/rtme-browser/vendor/github.com/sapk/go-genesys/db/rtme"
    }
  },
  "responses": {
    "BadRequest": {
      "description": "BadRequest is an bad request error response",
      "schema": {
        "$ref": "#/definitions/Error"
      }
    },
    "Error": {
      "description": "Error is error format response",
      "headers": {
        "error": {
          "type": "string"
        }
      }
    },
    "ErrorNotFound": {
      "description": "ErrorNotFound is a not found format response",
      "schema": {
        "$ref": "#/definitions/Error"
      }
    },
    "ErrorUnauthorized": {
      "description": "ErrorUnauthorized is an unauthorized error response",
      "schema": {
        "$ref": "#/definitions/Error"
      }
    },
    "InternalServerError": {
      "description": "InternalServerError is a InternalServerError format response",
      "schema": {
        "$ref": "#/definitions/Error"
      }
    },
    "RTMELoginResponse": {
      "description": "RTMELoginResponse is a RTMELoginResponse format response",
      "schema": {
        "$ref": "#/definitions/FormattedLoginResp"
      }
    },
    "RTMEStatusResponse": {
      "description": "RTMEStatusResponse is a RTMEStatusResponse format response",
      "schema": {
        "$ref": "#/definitions/FormattedStatusResp"
      }
    }
  }
}